Support fullnameOverride
This enables Rancher to be packaged as a subchart of an umbrella chart,
per CSP marketplace requirement. Though the umbrella chart will still
need to use the cattle-system namespace, the service account
and role binding resource names can now be overwritten with
fullnameOverride, which should not clash with other subcharts under the
same umbrella chart.